pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "djai"
version = "0.2.0"
description = "Utilities for working with Django-based AI integrations."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "Your Name", email = "you@example.com" },
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/djai"
Repository = "https://github.com/your-org/djai"
Issues = "https://github.com/your-org/djai/issues"

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.5.0",
]

[project.scripts]
djai = "djai.cli:main"

[tool.hatch.version]
path = "src/djai/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/djai"]

[tool.hatch.build.targets.sdist]
include = [
  "src/djai",
  "README.md",
  "LICENSE",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"


